move environment.h/cc from framework to session project/folder.
