+++ onnxruntime/server/executor.cc
@@ -5,7 +5,7 @@
 #include <onnx/onnx_pb.h>
 #include "core/common/logging/logging.h"
 #include "core/framework/data_types.h"
-#include "core/framework/environment.h"
+#include "core/session/environment.h"
 #include "core/framework/framework_common.h"
 #include "core/framework/mem_buffer.h"
 #include "core/framework/ml_value.h"
