+++ onnxruntime/test/framework/test_main.cc
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-#include "core/framework/environment.h"
+#include "core/session/environment.h"
 #include "gtest/gtest.h"
 #include "test/test_environment.h"
